Revising the github.client-services as well as restructring the layout
