+++ next.config.js
@@ -4,11 +4,11 @@ const nextConfig = {
   basePath: process.env.NODE_ENV === 'production' ? '/openbc-web' : '',
   output: 'export',
   distDir: 'dist',
+  serverRuntimeConfig: {
+    githubApiToken: process.env.GITHUB_API_KEY,
+  },
   images: {
     unoptimized: true,
-  },
-  env:{
-    GITHUB_API_KEY: process.env.GITHUB_TOKEN,
   },
 }
